fix: move raw and analytics files to dirs, change go version to 1.24.12
