+++ go.mod
@@ -1,5 +1,5 @@
 module github.com/Go-Yadro-Group-1
 
-go 1.22.7
+go 1.24.12
 
 require github.com/lib/pq v1.11.1
