+++ repository/analytics/activity_by_task.go
@@ -1,4 +1,4 @@
-package repository
+package analytics
 
 import (
 	"context"
