+++ repository/analytics/analytics_repository.go
@@ -1,4 +1,4 @@
-package repository
+package analytics
 
 import (
 	"context"
