+++ repository/analytics/open_task_time.go
@@ -1,4 +1,4 @@
-package repository
+package analytics
 
 import (
 	"context"
