+++ repository/analytics/task_priority_count.go
@@ -1,4 +1,4 @@
-package repository
+package analytics
 
 import (
 	"context"
