+++ repository/raw/issue.go
@@ -1,4 +1,4 @@
-package repository
+package raw
 
 import (
 	"context"
